+++ CHANGELOG.md
@@ -1,7 +1,12 @@
+# 1.1.2
+### Fixes
+- Tweens fast-forwarding many loops in a single update will call `onLoop` as many times as loops advanced.
+  - Added a test for this.
+
 # 1.1.1
 ### Fixes
-- Calling `end` on an infinite looping tween will not cause a recursive endless loop.
-  - Added a test for this. Needs a better way to check for endless looping but will do for now.
+- Calling `end` on an infinite repeat tween will not cause a recursive endless loop.
+  - Added a test for this. Needs a better way to check for endless repeat but will do for now.
 - Calling `end` on a repeating tween will fast-forward it to the end of all the loops and call the `onComplete`.
   - Extended the test on really big delta times for this since `end` is an infinite delta time update.